requirements.txt
# DOC ANOMALY DETECTION SYSTEM - Requirements

# Core ML/AI Frameworks
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# OpenAI
openai>=1.3.0

# AWS
boto3>=1.28.0

# Streamlit
streamlit>=1.28.0
streamlit-aggrid>=0.3.0
streamlit-plotly-events>=0.0.6

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=10.0.0
pytesseract>=0.3.10

# Text Processing
nltk>=3.8.0
spacy>=3.6.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0
regex>=2023.0.0
pyyaml>=6.0.0

# Reinforcement Learning
stable-baselines3>=2.2.0
gymnasium>=0.29.0
